String format specifier warnings.

+++ MultiplePageView.m
@@ -220,7 +220,7 @@ static float defaultTextPadding(void) {
 	[pageStorage addLayoutManager:lm];
 	 // owned by the text storage
 	
-	NSAttributedString *formfeed = [[NSAttributedString alloc] initWithString:[NSString stringWithFormat:@"%C", NSFormFeedCharacter] attributes:nil];
+	NSAttributedString *formfeed = [[NSAttributedString alloc] initWithString:[NSString stringWithFormat:@"%C", (unsigned short)NSFormFeedCharacter] attributes:nil];
 	NSFont *bodyFont = [[GlobalPrefs defaultPrefs] noteBodyFont];
 	
 	unsigned i, totalPageCount = 0; //[tableView numberOfSelectedRows];
